requirements.txt
django==4.1
djangorestframework
markdown
django-filter

celery[redis]

psycopg2-binary
